Image_to_Text/requirements.txt
python-dotenv
transformers 